xilinx usb2 gadget: get rid of incredibly annoying compile warning
commit 7796c11c728ad40ba4151d559a949c002deffb9a upstream.

This one was driving me mad, with several lines of warnings during the
allmodconfig build for a single bogus pointer cast.  The warning was so
verbose due to the indirect macro expansion explanation, and the whole
thing was just for a debug printout.

The bogus pointer-to-integer cast was pointless anyway, so just remove
it, and use '%p' to show the pointer.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

# Conflicts:
#	drivers/usb/gadget/udc/udc-xilinx.c